Payroll: Split tests into separate files
